#1 [refactor]: NLP 기능 통합 및 인증 연동
