+++ fastapi-server/requirements.txt
@@ -5,3 +5,4 @@ httpx==0.26.0
 python-dotenv==1.0.0
 pydantic==2.5.3
 redis==5.0.1
+konlpy==0.6.0
