--- nlp-service/.dockerignore
@@ -1,6 +1,0 @@
-.venv
-__pycache__
-*.pyc
-*.pyo
-*.pyd
-.Python
