--- nlp-service/Dockerfile
@@ -1,22 +1,0 @@
-FROM python:3.11-slim
-
-WORKDIR /app
-
-# 시스템 의존성 설치 (KoNLPy용 Java)
-RUN apt-get update && apt-get install -y \
-    default-jdk \
-    curl \
-    && rm -rf /var/lib/apt/lists/*
-
-# Python 의존성 설치
-COPY requirements.txt .
-RUN pip install --no-cache-dir -r requirements.txt
-
-# 애플리케이션 코드 복사
-COPY main.py .
-
-# 포트 노출
-EXPOSE 8000
-
-# FastAPI 서버 실행
-CMD ["uvicorn", "main:app", "--host", "0.0.0.0", "--port", "8000"]
