--- nlp-service/requirements.txt
@@ -1,4 +1,0 @@
-fastapi==0.104.1
-uvicorn==0.24.0
-pydantic==2.5.0
-konlpy==0.6.0
